organize in subpackages and add explanatory comments
