+++ DatabindingWithRecyclerView/app/src/androidTest/java/com/enpassio/databindingwithrecyclerview/ExampleInstrumentedTest.java
@@ -1,13 +1,13 @@
 package com.enpassio.databindingwithrecyclerview;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Instrumented test, which will execute on an Android device.
