+++ DatabindingWithRecyclerView/app/src/main/java/com/enpassio/databindingwithrecyclerview/ui/DetailsFragment.java
@@ -1,4 +1,4 @@
-package com.enpassio.databindingwithrecyclerview;
+package com.enpassio.databindingwithrecyclerview.ui;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.enpassio.databindingwithrecyclerview.model.Product;
+import com.enpassio.databindingwithrecyclerview.R;
 import com.enpassio.databindingwithrecyclerview.databinding.FragmentDetailsBinding;
 
 import androidx.annotation.NonNull;
@@ -14,7 +16,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.databinding.DataBindingUtil;
 import androidx.fragment.app.Fragment;
 
-import static com.enpassio.databindingwithrecyclerview.ProductListFragment.PRODUCT_KEY;
+import static com.enpassio.databindingwithrecyclerview.ui.ProductListFragment.PRODUCT_KEY;
 
 public class DetailsFragment extends Fragment {
 
@@ -27,12 +29,14 @@ public class DetailsFragment extends Fragment {
         FragmentDetailsBinding binding = DataBindingUtil.inflate(
                 inflater, R.layout.fragment_details, container, false);
 
+        //These are for making up button work. Not related to databinding.
         ((AppCompatActivity)getActivity()).setSupportActionBar(binding.toolbar);
-
         setHasOptionsMenu(true);
 
         Bundle bundle = getArguments();
         if (bundle != null) {
+            /*Once we get the chosen product from the bundle,
+            we pass it to the binding implementation*/
             Product chosenProduct = bundle.getParcelable(PRODUCT_KEY);
             binding.setProduct(chosenProduct);
         }
@@ -42,6 +46,7 @@ public class DetailsFragment extends Fragment {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
+        //This is for making up button in the toolbar behave like back button
         if(item.getItemId() == android.R.id.home){
             getFragmentManager().popBackStack();
         }
